pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "s1-snowdepth"
version = "0.1.0"
description = "Tools for operationalizing Dr. Dunmire's machine learning model to estimate snow depth across the European Alps using Sentinel-1 SAR backscatter."
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pandas",
    "xarray",
    "rioxarray",
    "rasterio",
    "scipy",
    "pycaret[full]",
    "shap",
    "python-dotenv",
    "asf_search",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]

